requirement.txt
ansible==2.6.4
Django==1.11.8
django-pure-pagination
PyMySQL==0.9.2
django-celery==3.2.2
redis==2.10.6
celery-with-redis==3.0 
python-gitlab==1.6.0 
python-jenkins==1.2.1
zabbix-client==0.1.1
influxdb==5.2.0
requests
